+++ lwutil/src/include/lwutil/lwutil.h
@@ -29,7 +29,7 @@
  * This file is part of LwUTIL - Lightweight utility library.
  *
  * Author:          Tilen MAJERLE <tilen@majerle.eu>
- * Version:         v1.3.0
+ * Version:         v1.4.0
  */
 #ifndef LWUTIL_HDR_H
 #define LWUTIL_HDR_H
@@ -332,6 +332,130 @@ lwutil_ld_u16_be(const void* ptr) {
 static inline uint32_t
 lwutil_ld_u32_be(const void* ptr) {
     const uint8_t* p = (const uint8_t*)ptr;
+    return p[0] << 24 | p[1] << 16 | p[2] << 8 | p[3];
+}
+
+/**
+ * \brief           Store `16-bit` value to bytes array in little-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       val: Value to write to output array
+ * \param[out]      ptr: Minimum `2-bytes` long output array to write value to
+ */
+static inline void
+lwutil_st_u16_le_ex(uint16_t val, void** ptr) {
+    uint8_t* p = (uint8_t*)(*ptr);
+
+    p[0] = (uint8_t)((val >> 0) & 0xFF);
+    p[1] = (uint8_t)((val >> 8) & 0xFF);
+    *ptr = (uint8_t*)(*ptr) + 2;
+}
+
+/**
+ * \brief           Store `32-bit` value to bytes array in little-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       val: Value to write to output array
+ * \param[out]      ptr: Minimum `4-bytes` long output array to write value to
+ */
+static inline void
+lwutil_st_u32_le_ex(uint32_t val, void** ptr) {
+    uint8_t* p = (uint8_t*)(*ptr);
+
+    p[0] = (uint8_t)((val >> 0) & 0xFF);
+    p[1] = (uint8_t)((val >> 8) & 0xFF);
+    p[2] = (uint8_t)((val >> 16) & 0xFF);
+    p[3] = (uint8_t)((val >> 24) & 0xFF);
+    *ptr = (uint8_t*)(*ptr) + 4;
+}
+
+/**
+ * \brief           Load `16-bit` value from bytes array in little-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       ptr: Minimum `2-bytes` long input array to extract bytes from
+ * \return          `16-bit` value extracted from input array
+ */
+static inline uint16_t
+lwutil_ld_u16_le_ex(const void** ptr) {
+    const uint8_t* p = (const uint8_t*)(*ptr);
+    *ptr = (uint8_t*)(*ptr) + 2;
+    return p[1] << 8 | p[0];
+}
+
+/**
+ * \brief           Load `32-bit` value from bytes array in little-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       ptr: Minimum `2-bytes` long input array to extract bytes from
+ * \return          `32-bit` value extracted from input array
+ */
+static inline uint32_t
+lwutil_ld_u32_le_ex(const void** ptr) {
+    const uint8_t* p = (const uint8_t*)(*ptr);
+    *ptr = (uint8_t*)(*ptr) + 4;
+    return p[3] << 24 | p[2] << 16 | p[1] << 8 | p[0];
+}
+
+/**
+ * \brief           Store `16-bit` value to bytes array in big-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       val: Value to write to output array
+ * \param[out]      ptr: Minimum `2-bytes` long output array to write value to
+ */
+static inline void
+lwutil_st_u16_be_ex(uint16_t val, void** ptr) {
+    uint8_t* p = (uint8_t*)(*ptr);
+
+    p[0] = (uint8_t)((val >> 8) & 0xFF);
+    p[1] = (uint8_t)((val >> 0) & 0xFF);
+    *ptr = (uint8_t*)(*ptr) + 2;
+}
+
+/**
+ * \brief           Store `32-bit` value to bytes array in big-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       val: Value to write to output array
+ * \param[out]      ptr: Minimum `4-bytes` long output array to write value to
+ */
+static inline void
+lwutil_st_u32_be_ex(uint32_t val, void** ptr) {
+    uint8_t* p = (uint8_t*)(*ptr);
+
+    p[0] = (uint8_t)((val >> 24) & 0xFF);
+    p[1] = (uint8_t)((val >> 16) & 0xFF);
+    p[2] = (uint8_t)((val >> 8) & 0xFF);
+    p[3] = (uint8_t)((val >> 0) & 0xFF);
+    *ptr = (uint8_t*)(*ptr) + 4;
+}
+
+/**
+ * \brief           Load `16-bit` value from bytes array in big-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       ptr: Minimum `2-bytes` long input array to extract bytes from
+ * \return          `16-bit` value extracted from input array
+ */
+static inline uint16_t
+lwutil_ld_u16_be_ex(const void** ptr) {
+    const uint8_t* p = (const uint8_t*)(*ptr);
+    *ptr = (uint8_t*)(*ptr) + 2;
+    return p[0] << 8 | p[1];
+}
+
+/**
+ * \brief           Load `32-bit` value from bytes array in big-endian format
+ * \note            Extended version, accepts the pointer to pointer and modifies the address pointer points to.
+ *                  Input pointer will point after the used bytes
+ * \param[in]       ptr: Minimum `4-bytes` long input array to extract bytes from
+ * \return          `32-bit` value extracted from input array
+ */
+static inline uint32_t
+lwutil_ld_u32_be_ex(const void** ptr) {
+    const uint8_t* p = (const uint8_t*)(*ptr);
+    *ptr = (uint8_t*)(*ptr) + 4;
     return p[0] << 24 | p[1] << 16 | p[2] << 8 | p[3];
 }